Some Refactoring
