+++ mopidy_tidal/playback.py
@@ -1,10 +1,10 @@
 from __future__ import unicode_literals
 
-import inspect
 import logging
 
 from mopidy import backend
 from mopidy_tidal.lru_cache import track_cache
+from mopidy_tidal.utils import inspect_stack
 
 logger = logging.getLogger(__name__)
 
@@ -13,13 +13,13 @@ class TidalPlaybackProvider(backend.PlaybackProvider):
 
     def translate_uri(self, uri):
         logger.info("TIDAL uri: %s", uri)
-        logger.debug(''.join('{}\n{}'.format(':'.join(str(x) for x in i[1:-2]), ''.join(i[-2]))
-                            for i in inspect.stack()))
+        logger.debug(inspect_stack())
         parts = uri.split(':')
         track_id = int(parts[4])
         newurl = self.backend.session.get_media_url(track_id)
-        logger.info("transformed into %s", newurl)
         cached_track = track_cache.hit(uri)
-        logger.info('Track transformed: %s', "({t.length}) {t.name} {t.artists!s} {t.album!s}".format(t=cached_track)
-                    if cached_track else "Not cached")
+        logger.info("TIDAL track: ({t.length}) {t.name} - {artists} : {t.album.name}".format(
+            t=cached_track, artists=' & '.join(a.name for a in cached_track.artists))
+            if cached_track else "Not cached")
+        logger.info("translated into %s", newurl)
         return newurl
